Separate Relocation.h
